+++ src/index.ts
@@ -1,0 +1,5 @@
+export { KafkaProducer } from './lib/kafkaHighLevelProducer';
+export { KafkaConsumer } from './lib/kafkaConsumer';
+
+export { KafkaConsumerMock } from './mock/kafkaConsumer.mock';
+export { KafkaProducerMock } from './mock/kafkaProducer.mock';